airflow/requirements.txt
psycopg2-binary==2.9.9
apache-airflow==2.9.2
boto3
polars
adbc_driver_manager
adbc-driver-postgresql
pyarrow
